+++ packages/react-ui/src/widgets/SkeletonWidgets.js
@@ -57,3 +57,7 @@ export const SkeletonThinBarItem = styled(Skeleton)(({ theme }) => ({
 export const SkeletonMask = styled(Skeleton)(({ theme }) => ({
   backgroundColor: theme.palette.common.white
 }));
+
+export const SkeletonSolid = styled(Skeleton)(({ theme }) => ({
+  backgroundColor: theme.palette.grey[100]
+}));