+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Not released
 
+## 2.0
+
+### 2.0.5 (2023-05-26)
+
 - ScatterPlot Widget: Add a skeleton for loading state [#690](https://github.com/CartoDB/carto-react/pull/690)
 - Table Widget: Add a skeleton for loading state [#689](https://github.com/CartoDB/carto-react/pull/689)
 - TimeSeries Widget: Add a skeleton for loading state [#686](https://github.com/CartoDB/carto-react/pull/686)
@@ -13,8 +17,6 @@
 - Category & ComparativeCategory Widgets: Add a skeleton for loading state [#679](https://github.com/CartoDB/carto-react/pull/679)
 - Bar & Histogram & Formula & ComparativeFormula Widgets: Add a skeleton for loading state [#674](https://github.com/CartoDB/carto-react/pull/674)
 - Improve Source types [#687](https://github.com/CartoDB/carto-react/pull/687)
-
-## 2.0
 
 ### 2.0.4 (2023-05-19)